Iteration24 (#659)
* #590 : implementing functional test
* #620 : introducing out ref variable
* refactoring code - using out ref parameters
* warn if out variable is not assigned in the method
* [FIXED] ?. operator
* working on ide for linux
* [ADDED] #655 : new option to keep output terminal after the program ends
* [FIXED] retrieving a source for the reference modules
* [ADDED] auto-recompile option
* [FIXED] MTA : critical exception handler
* [FIXED] recent list - only 9 elements must be in the list
* [ADDED]#69 : GC statistics
* [ADDED] support project collection file
* [FIXED] nil cannot be returned for nil object - use default instead
* #650 : Calling a named constructor from an implicit default one

---------

+++ elenasrc3/gui/text.h
@@ -244,7 +244,7 @@ namespace elena_lang
          _length = NOTFOUND_POS;
       }
 
-      bool isEOF() const { return _page.last() && _offset >= (int)(*_page).used; }
+      bool isEOF() const { return _page.last() && _offset >= (*_page).used; }
       bool isEOL() { return length_pos() <= (pos_t)_column; }
 
       TextBookmark();
@@ -292,7 +292,7 @@ namespace elena_lang
       void insert(TextBookmark bookmark, const_text_t s, size_t length, bool checkRowCount);
       void erase(TextBookmark bookmark, size_t length, bool checkRowCount);
 
-      bool compare(TextBookmark bookmark, const_text_t line, size_t len, bool matchCase, 
+      bool compare(TextBookmark bookmark, const_text_t line, size_t len, bool matchCase,
          const_text_t terminators);
 
    public:
